requirements.txt
selenium==4.6.0
webdriver_manager==3.8.4
Flask-restful==0.3.9